+++ test/smokes/checkstyle/default_rules/src/Hello.java
@@ -1,0 +1,12 @@
+package sample.checkstyle;
+
+public class Main {
+
+    /**
+     * This is a sample program.
+     * @param args the command-line arguments
+     */
+    public static void main(String[] args) {
+        System.out.println("Hello Checkstyle!!");
+    }
+}