+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Misc:
 - **remark-lint** Add file path to error message [#2067](https://github.com/sider/runners/pull/2067)
 - **remark-lint** Add rule ID to error message [#2076](https://github.com/sider/runners/pull/2076)
 - **remark-lint** Bump remark-preset-lint-sider from 0.6.0 to 0.7.0 (**breaking**) [#2078](https://github.com/sider/runners/pull/2078)
+- **Checkstyle** Provide Sider's recommended ruleset [#1619](https://github.com/sider/runners/pull/1619)
 
 ## 0.42.1